refactor: user 내 enum 타입을 통한 필드값 구분
