+++ backend/footprint/src/main/java/foot/footprint/domain/user/AuthProvider.java
@@ -1,0 +1,5 @@
+package foot.footprint.domain.user;
+
+public enum AuthProvider {
+    local, google, naver, kakao
+}
