+++ backend/footprint/src/main/java/foot/footprint/domain/user/Role.java
@@ -1,0 +1,5 @@
+package foot.footprint.domain.user;
+
+public enum Role {
+    admin, user
+}
